Separate implementation of handler and some refactoring
