--- src/main/java/me/uniodex/velocityrcon/commandsource/RconCommandSource.java
@@ -1,5 +1,0 @@
-package me.uniodex.velocityrcon.commandsource;
-
-import com.velocitypowered.api.command.CommandSource;
-
-public interface RconCommandSource extends CommandSource {}
