+++ src/main/java/me/uniodex/velocityrcon/server/RconServer.java
@@ -1,6 +1,5 @@
 package me.uniodex.velocityrcon.server;
 
-import com.velocitypowered.api.proxy.ProxyServer;
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelInitializer;
@@ -10,47 +9,46 @@ import io.netty.channel.socket.SocketChannel;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
 import lombok.Getter;
 
-import java.net.SocketAddress;
+import java.net.InetSocketAddress;
 
 public class RconServer {
 
     @Getter
-    private final ProxyServer server;
+    private final RconHandler handler;
 
-    private ServerBootstrap bootstrap = new ServerBootstrap();
+    @Getter final InetSocketAddress address;
+
+    private final ServerBootstrap bootstrap = new ServerBootstrap();
     private final EventLoopGroup bossGroup = new NioEventLoopGroup();
     private final EventLoopGroup workerGroup = new NioEventLoopGroup();
 
-    public RconServer(ProxyServer server, final String password) {
-        this.server = server;
+    public RconServer(InetSocketAddress address, final String password, RconHandler handler) {
+        this.handler = handler;
+        this.address = address;
 
         bootstrap
                 .group(bossGroup, workerGroup)
                 .channel(NioServerSocketChannel.class)
                 .childHandler(new ChannelInitializer<SocketChannel>() {
                     @Override
-                    public void initChannel(SocketChannel ch) throws Exception {
+                    public void initChannel(SocketChannel ch) {
                         ch.pipeline()
                                 .addLast(new RconFramingHandler())
-                                .addLast(new RconHandler(RconServer.this, password));
+                                .addLast(new RconChannelHandler(RconServer.this, password));
                     }
                 });
     }
 
-    /**
-     * Bind the server on the specified address.
-     *
-     * @param address The address.
-     * @return Netty channel future for bind operation.
-     */
-    public ChannelFuture bind(final SocketAddress address) {
+    public ChannelFuture bind() {
+        handler.onBind(this);
         return bootstrap.bind(address);
     }
 
     /**
-     * Shut the Rcon server down.
+     * Shut the RCON server down
      */
     public void shutdown() {
+        handler.onShutdown(this);
         workerGroup.shutdownGracefully();
         bossGroup.shutdownGracefully();
     }
